requirements.txt
fsspec>=2023.1.0
joblib>=1.0.0
matplotlib>=3.5.0
numpy>=1.21.0
openai>=1.0.0
opencv-python>=4.5.0
pandas>=1.5.0
pillow>=9.0.0
pyarrow>=10.0.0
python-dotenv>=1.0.0
scikit-image>=0.19.0
scikit-learn>=1.0.0
seaborn>=0.12.0
torch>=2.0.0
torchvision>=0.15.0
tqdm>=4.64.0
transformers>=4.30.0
yt-dlp>=2025.9.26
